Artigos feitos

+++ artigos/views.py
@@ -31,6 +31,9 @@ def novo_artigo_view(request):
 def edita_artigo_view(request, artigo_id):
     artigo = Artigo.objects.get(id=artigo_id)
 
+    if request.user != artigo.autor:
+        return redirect("artigos")
+
     if request.POST:
         form = ArtigoForm(request.POST or None, request.FILES, instance=artigo)
 
@@ -48,5 +51,79 @@ def edita_artigo_view(request, artigo_id):
 def apaga_artigo_view(request, artigo_id):
 
     artigo = Artigo.objects.get(id=artigo_id)
+
+    if request.user != artigo.autor:
+        return redirect("artigos")
+    
     artigo.delete()
+    return redirect("artigos")
+
+
+@login_required
+def like(request, artigo_id):
+    artigo = Artigo.objects.get(id=artigo_id)
+
+    if request.user in artigo.likes.all():
+        artigo.likes.remove(request.user)
+    else:
+        artigo.likes.add(request.user)
+
+    return redirect("artigos")
+
+
+@login_required
+def comentar_artigo(request, artigo_id):
+    artigo = Artigo.objects.get(id=artigo_id)
+
+    comentarios = artigo.comentarios.all()
+
+    form = ComentarioForm()
+
+    if request.POST:
+        if request.user.is_authenticated:
+            form = ComentarioForm(request.POST)
+
+            if form.is_valid():
+                comentario = form.save(commit=False)
+
+                comentario.artigo = artigo
+                comentario.autor = request.user
+
+                comentario.save()
+
+                return redirect("artigos")
+            
+    context = {"artigo":artigo, "comentarios": comentarios, "form":form}
+
+    return render(request, "artigos/comentar_artigo.html", context)
+
+
+@login_required
+def editar_comentario(request, comentario_id):
+    comentario = Comentario.objects.get(id=comentario_id)
+
+    if request.user != comentario.autor:
+        return redirect("artigos")
+
+    if request.POST:
+        form = ComentarioForm(request.POST or None, request.FILES, instance = comentario)
+
+        if form.is_valid():
+            form.save()
+            return redirect("artigos")
+    else:
+        form = ComentarioForm(instance=comentario)
+
+        context = {"form":form, "comentario":comentario}
+        return render(request, "artigos/edita_comentario.html", context)
+
+
+@login_required
+def apagar_comentario(request, comentario_id):
+    comentario = Comentario.objects.get(id=comentario_id)
+
+    if request.user != comentario.autor:
+        return redirect("artigos")
+
+    comentario.delete()
     return redirect("artigos")
